python/pyproject.toml
[tool.poetry]
name = "standard-bom"
version = "1.0.0"
description = "Standard BOM Format Library"
authors = ["Hakan Dilek <hakan.dilek@siemens.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "standardbom"}]

[tool.poetry.dependencies]
python = "^3.8.1"
cyclonedx-bom = "^3.7.2"
cyclonedx-python-lib = "^3.1.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
coverage = "^6.5.0"
pytest = "^7.2.0"
flake8 = "^6.0.0"
deepdiff = "^6.2.1"
mypy = "^0.991"
types-python-dateutil = "^2.8.19.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
